# plan_vision_backend/api-docs.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/sign-in/email": {
      "post": {
        "operationId": "AuthDocsController_signIn",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User signed in successfully"
          }
        },
        "summary": "Sign in with email and password",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/sign-up/email": {
      "post": {
        "operationId": "AuthDocsController_signUp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User signed up successfully"
          }
        },
        "summary": "Sign up with email and password",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/image-types": {
      "post": {
        "operationId": "ImageTypeController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateImageTypeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Image type created."
          }
        },
        "summary": "Create a new image type (admin use).",
        "tags": [
          "Image Types"
        ]
      },
      "get": {
        "operationId": "ImageTypeController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of image types."
          }
        },
        "summary": "List all image types sorted by label.",
        "tags": [
          "Image Types"
        ]
      }
    },
    "/image-types/{id}": {
      "get": {
        "operationId": "ImageTypeController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Image type details."
          },
          "404": {
            "description": "Image type not found."
          }
        },
        "summary": "Get a single image type by ID.",
        "tags": [
          "Image Types"
        ]
      },
      "patch": {
        "operationId": "ImageTypeController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateImageTypeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Image type updated."
          },
          "404": {
            "description": "Image type not found."
          }
        },
        "summary": "Update label/value/description.",
        "tags": [
          "Image Types"
        ]
      },
      "delete": {
        "operationId": "ImageTypeController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Image type deleted."
          },
          "404": {
            "description": "Image type not found."
          }
        },
        "summary": "Delete an image type.",
        "tags": [
          "Image Types"
        ]
      }
    },
...
    "/input-images": {
      "post": {
        "operationId": "InputImageController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInputImageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Image registered successfully."
          },
          "401": {
            "description": "Unauthorized. Session is required."
          }
        },
        "security": [
          {
            "access-token": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Register a new uploaded image URL for the current user.",
        "tags": [
          "Input Images"
        ]
      },
      "get": {
        "operationId": "InputImageController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of input images."
          },
          "401": {
            "description": "Unauthorized. Session is required."
          }
        },
        "security": [
          {
            "access-token": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "List input images for the current user.",
        "tags": [
          "Input Images"
        ]
      }
    },
    "/input-images/upload": {
      "post": {
        "operationId": "InputImageController_upload",
        "parameters": [],
        "responses": {
          "201": {
            "description": "File uploaded and registered successfully."
          },
          "400": {
            "description": "Bad Request. File is missing."
          },
          "401": {
            "description": "Unauthorized. Session is required."
          }
        },
        "security": [
          {
            "access-token": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Upload a file to Google Cloud Storage and register it.",
        "tags": [
          "Input Images"
        ]
      }
    },
    "/input-images/{id}/url": {
      "get": {
        "operationId": "InputImageController_getSignedUrl",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Signed URL returned."
          },
          "401": {
            "description": "Unauthorized. Session is required."
          },
          "404": {
            "description": "Image not found or does not belong to user."
          }
        },
        "security": [
          {
            "access-token": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Generate a signed URL for downloading the image.",
        "tags": [
          "Input Images"
        ]
      }
    },
    "/input-images/{id}": {
      "get": {
        "operationId": "InputImageController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Image details."
          },
          "401": {
            "description": "Unauthorized. Session is required."
          },
          "404": {
            "description": "Image not found or does not belong to user."
          }
        },
        "security": [
          {
            "access-token": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Get a single uploaded image by ID (must belong to user).",
        "tags": [
          "Input Images"
        ]
      },
      "delete": {
        "operationId": "InputImageController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Image record deleted."
          },
          "401": {
            "description": "Unauthorized. Session is required."
          },
          "404": {
            "description": "Image not found or does not belong to user."
          }
        },
        "security": [
          {
            "access-token": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Delete an uploaded image record.",
        "tags": [
          "Input Images"
        ]
      }
    },
...
  },
  "info": {
    "title": "Plan Vision API",
    "description": "REST API for Plan Vision backend (users, auth, providers).",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "auth_session",
        "description": "Better Auth session cookie"
      },
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Optional Better Auth access token"
      }
    },
    "schemas": {
      "SignInDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "User email address"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "description": "User password"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "SignUpDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John Doe",
            "description": "User full name"
          },
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "User email address"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "description": "User password"
          }
        },
        "required": [
          "name",
          "email",
          "password"
        ]
      },
      "CreateImageTypeDto": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "description": "User-facing label (e.g., \"2D Floor Plan\").",
            "maxLength": 255
          },
          "value": {
            "type": "string",
            "description": "Internal value key (e.g., \"floor_plan_2d\").",
            "maxLength": 255
          },
          "description": {
            "type": "string",
            "description": "Optional description shown in the UI."
          }
        },
        "required": [
          "label",
          "value"
        ]
      },
      "UpdateImageTypeDto": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "description": "User-facing label (e.g., \"2D Floor Plan\").",
            "maxLength": 255
          },
          "value": {
            "type": "string",
            "description": "Internal value key (e.g., \"floor_plan_2d\").",
            "maxLength": 255
          },
          "description": {
            "type": "string",
            "description": "Optional description shown in the UI."
          }
        }
      },
...
      "CreateInputImageDto": {
        "type": "object",
        "properties": {
          "imageUrl": {
            "type": "string",
            "description": "Cloud storage URL pointing to the uploaded file.",
            "example": "https://storage.googleapis.com/plan-vision/uploads/123.png"
          },
          "filename": {
            "type": "string",
            "description": "Original filename as provided by the user.",
            "example": "floor-plan.png"
          }
        },
        "required": [
          "imageUrl",
          "filename"
        ]
      },
...
    }
  }
}